reordered steps and updated labels
