+++ app/react/V2/Routes/Settings/ParagraphExtraction/components/extractors/CreateDialog/steps/ExtractionConfiguration/Footer.tsx
@@ -1,35 +1,76 @@
+//////
+
 import React from 'react';
 import { Button } from 'V2/Components/UI';
 import { Translate } from 'app/I18N';
-import { ArrowRightIcon } from '@heroicons/react/20/solid';
+import { isClient } from 'app/utils';
+import { captureException } from '@sentry/react';
+import * as extractorsAPI from 'V2/api/paragraphExtractor/extractors';
+import { notificationAtom } from 'V2/atoms';
+import { useRevalidator } from 'react-router';
+import { useSetAtom } from 'jotai';
 import { useCreateExtractorContext } from '../../CreateExtractorContext';
 
 const Footer = () => {
+  const revalidator = useRevalidator();
+  const setNotifications = useSetAtom(notificationAtom);
   const {
+    sourceTemplateId,
     setStep,
+    targetTemplateId,
+    setShowModal,
     paragraphPropertyId,
     paragraphNumberPropertyId,
     targetRelationshipId,
     sourceRelationshipId,
   } = useCreateExtractorContext();
+
   const isDisabled =
     !paragraphPropertyId ||
     !paragraphNumberPropertyId ||
     !targetRelationshipId ||
-    !sourceRelationshipId;
+    !sourceRelationshipId ||
+    !sourceTemplateId ||
+    !targetTemplateId;
+
+  const handleSubmit = async () => {
+    try {
+      const values = {
+        sourceTemplateId,
+        targetTemplateId,
+        paragraphPropertyId,
+        paragraphNumberPropertyId,
+        targetRelationshipId,
+        sourceRelationshipId,
+      };
+      await extractorsAPI.save(values);
+      setShowModal(false);
+      await revalidator.revalidate();
+      setNotifications({
+        type: 'success',
+        text: <Translate>Paragraph Extractor added</Translate>,
+      });
+    } catch (e) {
+      if (isClient) {
+        const error = new Error('Error saving extractor', { cause: e });
+        captureException(error);
+      }
+    }
+  };
+
   return (
     <>
-      <Button styling="light" onClick={() => setStep(1)} className="grow">
-        <Translate>Cancel</Translate>
+      <Button styling="light" onClick={() => setStep(2)} className="grow">
+        <Translate>Back</Translate>
       </Button>
       <Button
         className="grow disabled:opacity-50"
-        onClick={async () => setStep(3)}
+        color="success"
+        onClick={async () => handleSubmit()}
         disabled={isDisabled}
       >
         <span className="flex items-center justify-center gap-2 flex-nowrap">
-          <Translate>Next</Translate>
-          <ArrowRightIcon className="w-5" />
+          <Translate>Create</Translate>
         </span>
       </Button>
     </>
