+++ app/react/V2/Routes/Settings/ParagraphExtraction/components/extractors/CreateDialog/steps/SourceTemplate/Footer.tsx
@@ -1,66 +1,27 @@
 import React from 'react';
 import { Button } from 'V2/Components/UI';
 import { Translate } from 'app/I18N';
-import { isClient } from 'app/utils';
-import { captureException } from '@sentry/react';
-import * as extractorsAPI from 'V2/api/paragraphExtractor/extractors';
-import { notificationAtom } from 'V2/atoms';
-import { useRevalidator } from 'react-router';
-import { useSetAtom } from 'jotai';
+import { ArrowRightIcon } from '@heroicons/react/20/solid';
 import { useCreateExtractorContext } from '../../CreateExtractorContext';
+import { sourceTemplate } from 'api/paragraphExtraction/application/specs/fixtures';
 
 const Footer = () => {
-  const revalidator = useRevalidator();
-  const setNotifications = useSetAtom(notificationAtom);
-  const {
-    sourceTemplateId,
-    setStep,
-    targetTemplateId,
-    setShowModal,
-    paragraphPropertyId,
-    paragraphNumberPropertyId,
-    targetRelationshipId,
-    sourceRelationshipId,
-  } = useCreateExtractorContext();
-
-  const handleSubmit = async () => {
-    try {
-      const values = {
-        sourceTemplateId,
-        targetTemplateId,
-        paragraphPropertyId,
-        paragraphNumberPropertyId,
-        targetRelationshipId,
-        sourceRelationshipId,
-      };
-      await extractorsAPI.save(values);
-      setShowModal(false);
-      await revalidator.revalidate();
-      setNotifications({
-        type: 'success',
-        text: <Translate>Paragraph Extractor added</Translate>,
-      });
-    } catch (e) {
-      if (isClient) {
-        const error = new Error('Error saving extractor', { cause: e });
-        captureException(error);
-      }
-    }
-  };
+  const { setStep, sourceTemplateId } = useCreateExtractorContext();
+  const isDisabled = !sourceTemplateId;
 
   return (
     <>
-      <Button styling="light" onClick={() => setStep(2)} className="grow">
-        <Translate>Cancel</Translate>
+      <Button styling="light" onClick={() => setStep(1)} className="grow">
+        <Translate>Back</Translate>
       </Button>
       <Button
         className="grow disabled:opacity-50"
-        color="success"
-        onClick={async () => handleSubmit()}
-        disabled={!sourceTemplateId}
+        onClick={async () => setStep(3)}
+        disabled={isDisabled}
       >
         <span className="flex items-center justify-center gap-2 flex-nowrap">
-          <Translate>Create</Translate>
+          <Translate>Next</Translate>
+          <ArrowRightIcon className="w-5" />
         </span>
       </Button>
     </>
